Proje bitimi.

+++ VTDGP-donem-projesi-berdanaksoy/yoneticiDuzenlemeEkrani.cs
@@ -27,7 +27,7 @@ namespace VTDGP_donem_projesi_berdanaksoy
         {
             con = new SqlConnection("server=BERDAN\\SQLEXPRESS; Initial Catalog=VTDGP Proje Restaurant;Integrated Security=SSPI");
             con.Open();
-            da = new SqlDataAdapter("SELECT * FROM yoneticiGirisi", con);
+            da = new SqlDataAdapter("SELECT * FROM yoneticiGirisi order by yoneticiID", con);
             DataSet ds = new DataSet();
             da.Fill(ds);
             dataGridView1.DataSource = ds.Tables[0];
@@ -50,14 +50,16 @@ namespace VTDGP_donem_projesi_berdanaksoy
         {
             bool check = false;
             con = new SqlConnection("server=BERDAN\\SQLEXPRESS; Initial Catalog=VTDGP Proje Restaurant;Integrated Security=SSPI");
-            cmd = new SqlCommand("select sifre from yoneticiGirisi where yoneticiAdi=@yoneticiAdi", con);
+            cmd = new SqlCommand("select * from yoneticiGirisi where yoneticiAdi=@yoneticiAdi and sifre=@sifre", con);
             cmd.Parameters.AddWithValue("yoneticiAdi", textBox7.Text);
+            cmd.Parameters.AddWithValue("sifre",textBox2.Text);
+            con.Open();
             SqlDataReader dr = cmd.ExecuteReader();
             if (dr.Read())
             {
-                if (textBox1.Text == dr["sifre"].ToString())
+                if (textBox7.Text == dr["yoneticiAdi"].ToString() && textBox2.Text == dr["sifre"].ToString())
                 {
-                    check = true;
+                        check = true;
                 }
             }
             if (check==true)
@@ -69,59 +71,63 @@ namespace VTDGP_donem_projesi_berdanaksoy
                 cmd.Parameters.AddWithValue("sifre", textBox3.Text);
                 cmd.ExecuteNonQuery();
 
-                da = new SqlDataAdapter("SELECT * FROM yoneticiGirisi", con);
+                da = new SqlDataAdapter("SELECT * FROM yoneticiGirisi order by yoneticiID", con);
                 DataSet dss = new DataSet();
                 da.Fill(dss);
                 dataGridView1.DataSource = dss.Tables[0];
                 con.Close();
+
+                MessageBox.Show("Sifreniz degismistir.");
             }
             else
             {
-                MessageBox.Show("Sifreniz yanlis.");
+                MessageBox.Show("Yonetici adi veya sifreniz yanlis.");
             }
-            con = new SqlConnection("server=BERDAN\\SQLEXPRESS; Initial Catalog=VTDGP Proje Restaurant;Integrated Security=SSPI");
-            con.Open();
-            cmd = new SqlCommand("UPDATE yoneticiGirisi SET sifre = @sifre WHERE yoneticiAdi = @yoneticiAdi", con);
-            cmd.Parameters.AddWithValue("yoneticiAdi", textBox7.Text);
-            cmd.Parameters.AddWithValue("sifre", textBox3.Text);
-            cmd.ExecuteNonQuery();
-
-            da = new SqlDataAdapter("SELECT * FROM yoneticiGirisi", con);
-            DataSet ds = new DataSet();
-            da.Fill(ds);
-            dataGridView1.DataSource = ds.Tables[0];
-            con.Close();
         }
 
         private void button4_Click(object sender, EventArgs e)
         {
-            con = new SqlConnection("server=BERDAN\\SQLEXPRESS; Initial Catalog=VTDGP Proje Restaurant;Integrated Security=SSPI");
-            con.Open();
-            cmd = new SqlCommand("insert into yoneticiGirisi (yoneticiAdi,sifre) values (@yoneticiAdi,@sifre) ", con);
-            cmd.Parameters.AddWithValue("yoneticiAdi", textBox4.Text);
-            cmd.Parameters.AddWithValue("sifre", textBox5.Text);
-            cmd.ExecuteNonQuery();
+            if (textBox4.Text=="" && textBox5.Text=="")
+            {
+                MessageBox.Show("Yonetici adi ve sifre bos birakilamaz.");
+            }
+            else
+            {
+                con = new SqlConnection("server=BERDAN\\SQLEXPRESS; Initial Catalog=VTDGP Proje Restaurant;Integrated Security=SSPI");
+                con.Open();
+                cmd = new SqlCommand("insert into yoneticiGirisi (yoneticiAdi,sifre) values (@yoneticiAdi,@sifre) ", con);
+                cmd.Parameters.AddWithValue("yoneticiAdi", textBox4.Text);
+                cmd.Parameters.AddWithValue("sifre", textBox5.Text);
+                cmd.ExecuteNonQuery();
 
-            da = new SqlDataAdapter("SELECT * FROM yoneticiGirisi", con);
-            DataSet ds = new DataSet();
-            da.Fill(ds);
-            dataGridView1.DataSource = ds.Tables[0];
-            con.Close();
+                da = new SqlDataAdapter("SELECT * FROM yoneticiGirisi order by yoneticiID", con);
+                DataSet ds = new DataSet();
+                da.Fill(ds);
+                dataGridView1.DataSource = ds.Tables[0];
+                con.Close();
+            }
         }
 
         private void button5_Click(object sender, EventArgs e)
         {
-            con = new SqlConnection("server=BERDAN\\SQLEXPRESS; Initial Catalog=VTDGP Proje Restaurant;Integrated Security=SSPI");
-            con.Open();
-            cmd = new SqlCommand("delete from yoneticiGirisi where yoneticiID = @yoneticiID", con);
-            cmd.Parameters.AddWithValue("yoneticiID", textBox6.Text);
-            cmd.ExecuteNonQuery();
+            if (int.Parse(textBox6.Text)==1)
+            {
+                MessageBox.Show("Ana yonetici silinemez.");
+            }
+            else
+            {
+                con = new SqlConnection("server=BERDAN\\SQLEXPRESS; Initial Catalog=VTDGP Proje Restaurant;Integrated Security=SSPI");
+                con.Open();
+                cmd = new SqlCommand("delete from yoneticiGirisi where yoneticiID = @yoneticiID", con);
+                cmd.Parameters.AddWithValue("yoneticiID", textBox6.Text);
+                cmd.ExecuteNonQuery();
 
-            da = new SqlDataAdapter("SELECT * FROM yoneticiGirisi", con);
-            DataSet ds = new DataSet();
-            da.Fill(ds);
-            dataGridView1.DataSource = ds.Tables[0];
-            con.Close();
+                da = new SqlDataAdapter("SELECT * FROM yoneticiGirisi order by yoneticiID", con);
+                DataSet ds = new DataSet();
+                da.Fill(ds);
+                dataGridView1.DataSource = ds.Tables[0];
+                con.Close();
+            }
         }
 
         private void textBox6_KeyPress(object sender, KeyPressEventArgs e)
